requirements.txt
flask
celery
openai-whisper
psycopg2-binary
python-dotenv
SpeechRecognition
pyaudio
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-api-python-client
jinja2